Implement challenge 1, converting hex to base64
This turned out to be not that bad, once I realised how the different
bit sequences overlap:

In hex each character represents 4 bits (as 2^4 = 16), while in base64
each character represents 6 bits (2^6 = 64). This means that the
representation shrinks by a factor of 1.5 when converting, because we
can cram 1.5 times as many bits into every character.

Hex:
+----+----+----+
|  4 |  9 |  2 |
+----+----+----+

Binary:
0100 1001 0010
010010  010010

Base64:
+------+------+
|   S  |   S  |
+------+------+