Solve challenge 3
This one is about finding a good result from a set of results produced
by applying different decryption keys. I just pulled a distribution of
characters in the English language off the internet.

Because my distribution is case-insensitive, and does not include
punctuation or whitespace, I'm slightly fudging this to only use
lowercase characters as keys. I previously got an intelligible, but
malformed result when trying the whole ASCII range as keys.