Add a working solution for 1-6
This one took me a long time, because I did not understand the
implications of decoding base64 correctly. I got it correct earlier
when converting between hex and base64 directly, but forget what I
head learned when separating the bits.

Base64 has 6-bit characters, while we are working with 8-bit bytes.
That means to read in a base64 string as a Vec<u8>, we need to grab
four characters and make them into three bytes. Because encoded byte
sequences are not always neat multiples of four, there can be padding
at the end, which becomes annoying to en- and decode.

When decoding, we have to replace the padding with zeroes, so that the
chunking works out nicely, but then remove a bit off the end of
result.

I can highly recommend the base64 Wikipedia page, it is very well
done.