Implement the shift rows step for AES
This is by far the easiest one to implement.

In theory we could assign fields directly into the correct positions
rather than pushing and popping, but this is much shorter. We're
already using a mutable reference for performance purposes, to avoid
constructing a new nested vector from scratch.

To shave the yak even further, we could also use a regular vector
instead of a nested one, but again, this is clearer to the operator.