Implement the MixColumns step
This one had me puzzled for a long time because I just did not
understand what it was supposed to be doing. I ended up resorting to
the Wikipedia explanation, which has a code sample.

I actually still do not understand what it is doing, but I can study
the maths after committing this apparently working piece of code.