Implement decryption
This adds a lot of constant overhead. On the plus side, I finally
grokked how MixColumns works, and did away with the magical
implementation from Wikipedia. Essentially this is just doing matrix
multiplication with polynomials, but with XOR as addition. This is
surprisingly similar to algebraic normal form, which is used for
generating logic circuits from truth tables, among other things.

Because the inverse MixColumns step includes some higher-order
multiplications, it is easier to just use a static multiplication with
modulo table. Also, they were already on Wikipedia ready to copy.