Solve challenge 7, ECB AES-128 decryption
This took a few weeks because I had to implement AES first, which was
a good learning exercise.

The final tying together has shown how annoying my AES interface is,
which I should probably refactor to automatically handle the
transposing of in- and output data. Another pain point is the
construction of a key from a byte-array.

There are some empty bytes at the end of the decrypted string, likely
padding which I am not stripping properly, but I am not sure what kind
of padding it even is (AES? ECB? Something else?)